Factor out adding logic into handler
